+++ docs/src/examples/components/Text/Variations/TextExampleImportant.shorthand.tsx
@@ -1,0 +1,6 @@
+import React from 'react'
+import { Text } from '@stardust-ui/react'
+
+const TextExampleImportantShorthand = () => <Text important content="This text is important!" />
+
+export default TextExampleImportantShorthand