+++ docs/src/examples/components/Text/Variations/TextExampleImportant.tsx
@@ -1,0 +1,6 @@
+import React from 'react'
+import { Text } from '@stardust-ui/react'
+
+const TextExampleImportant = () => <Text important>This text is important!</Text>
+
+export default TextExampleImportant