+++ src/components/Text/textVariables.ts
@@ -1,0 +1,9 @@
+export interface ITextVariables {
+  importantWeight: number
+}
+
+export default (siteVars: any): ITextVariables => {
+  return {
+    importantWeight: 600,
+  }
+}